Fix warning
git-svn-id: https://svn.osgeo.org/gdal/trunk@40460 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_json_streaming_parser.cpp
@@ -862,7 +862,7 @@ std::string CPLJSonStreamingParser::GetSerializedString(const char* pszStr)
             osStr += "\\\"";
         else if( ch == '\\' )
             osStr += "\\\\";
-        else if( ch >= 0 && ch < ' ' )
+        else if( static_cast<int>(ch) >= 0 && ch < ' ' )
             osStr += CPLSPrintf("\\u%04X", ch);
         else
             osStr += ch;
